Add config to keep temp workdir files

+++ KindlerBot/Commands/ConvertCmd.cs
@@ -6,6 +6,7 @@ using KindlerBot.Configuration;
 using KindlerBot.Conversion;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 
@@ -16,21 +17,27 @@ namespace KindlerBot.Commands
     internal class ConvertCmdHandler : IRequestHandler<ConvertCmdRequest>
     {
         private readonly ITelegramBotClient _botClient;
-        private readonly IConfigStore _configManager;
+        private readonly IConfigStore _configStore;
         private readonly ICalibreCli _calibreCli;
+        private readonly ConversionConfiguration _conversionConfig;
         private readonly ILogger<ConvertCmdHandler> _logger;
 
-        public ConvertCmdHandler(ITelegramBotClient botClient, IConfigStore configManager, ICalibreCli calibreCli, ILogger<ConvertCmdHandler> logger)
+        public ConvertCmdHandler(ITelegramBotClient botClient,
+            IConfigStore configStore,
+            ICalibreCli calibreCli,
+            IOptions<ConversionConfiguration> conversionConfig,
+            ILogger<ConvertCmdHandler> logger)
         {
             _botClient = botClient;
-            _configManager = configManager;
+            _configStore = configStore;
             _calibreCli = calibreCli;
+            _conversionConfig = conversionConfig.Value;
             _logger = logger;
         }
 
         public async Task<Unit> Handle(ConvertCmdRequest request, CancellationToken cancellationToken)
         {
-            var email = await _configManager.GetChatEmail(request.Chat.Id);
+            var email = await _configStore.GetChatEmail(request.Chat.Id);
             if (email == null)
             {
                 await _botClient.SendTextMessageAsync(request.Chat, "❌ Email is not configured. Fix it and try again!", cancellationToken: cancellationToken);
@@ -48,6 +55,10 @@ namespace KindlerBot.Commands
             try
             {
                 using var tempDir = new TempDir(chat.Id);
+                if (_conversionConfig.KeepTempWorkDir)
+                {
+                    tempDir.SuppressCleanup();
+                }
 
                 await _botClient.SendTextMessageAsync(chat, "⏬ Downloading file");
                 var fileInfo = await _botClient.GetFileAsync(doc.FileId);
@@ -104,6 +115,7 @@ namespace KindlerBot.Commands
 
         private class TempDir: IDisposable
         {
+            private bool Clean { get; set; } = true;
             public string DirPath { get; }
 
             public TempDir(ChatId chatId)
@@ -112,9 +124,14 @@ namespace KindlerBot.Commands
                 Directory.CreateDirectory(DirPath);
             }
 
+            public void SuppressCleanup() => Clean = false;
+
             public void Dispose()
             {
-                Directory.Delete(DirPath, recursive: true);
+                if (Clean)
+                {
+                    Directory.Delete(DirPath, recursive: true);
+                }
             }
         }
     }
